% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_digits}
\alias{add_digits}
\title{Add leading digits in a vector}
\usage{
add_digits(x, digit = 0, seq_length)
}
\arguments{
\item{x}{A charecter vector.}

\item{digit}{The leading digit to be added.}

\item{seq_length}{required length of sequence of numbers.}
}
\description{
This function adds leading digits in a vector
}
\examples{
add_digits("(31) 9-8932-1010", 0, 12)
}
