% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{as_decade}
\alias{as_decade}
\title{Convert years into decades}
\usage{
as_decade(year)
}
\arguments{
\item{year}{numeric variable of years}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Returns a vector by grouping a string of years into decades
}
\examples{
as_decade(c(2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010, 2011))
}
