% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_r_pipe}
\alias{check_r_pipe}
\title{Check if the running R version contains the native pipe}
\usage{
check_r_pipe(.print = FALSE)
}
\arguments{
\item{.print}{prints the running R version}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function checks if the running R version contains the native pipe
}
