% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_var}
\alias{check_var}
\alias{check_element}
\title{Check if a named element exists within a \code{r} object}
\usage{
check_var(obj, x)

check_element(obj, x)
}
\arguments{
\item{obj}{\code{r} object}

\item{x}{named element on which existence will be checked}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function checks if a named element exists within a \code{r} object
}
\examples{
check_var(mtcars, cyl)

data_list = list(dat1 = mtcars, dat2 = airmiles)
check_element(data_list, dat1)
check_element(data_list, dat3)

}
