% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_dirnames}
\alias{clean_dirnames}
\title{Clean directory names}
\usage{
clean_dirnames(.path)
}
\arguments{
\item{.path}{path}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function cleans directory names
}
