% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_strings}
\alias{clean_strings}
\title{Clean string vectors}
\usage{
clean_strings(.var, .rm_accent = TRUE, .rm_spaces = TRUE)
}
\arguments{
\item{.var}{string vector}

\item{.rm_accent}{remove accents}

\item{.rm_spaces}{remove whitespace}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function cleans up string vectors, removing accents, whitespace, and
transforming characters from uppercase to lowercase.
}
