% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_duplicates}
\alias{collapse_duplicates}
\title{Collapse a data frame using group variables}
\usage{
collapse_duplicates(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{group variables}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function collapses a data frame using group variables
}
\note{
When data is collapsed, missing values are excluded.
}
