% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.files}
\alias{download.files}
\title{Download a set of files from the internet}
\usage{
download.files(x)
}
\arguments{
\item{x}{A character vector of urls.}
}
\description{
This function downloads a set of files from the internet.
The result is a named list with the tempfile path of each downloaded file.
}
