% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_numbers}
\alias{extract_numbers}
\title{Extract numerical sequences from strings}
\usage{
extract_numbers(.var)
}
\arguments{
\item{.var}{variable of type 'character'´}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function extracts numerical sequences from strings
}
