% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_merged_cells}
\alias{fill_merged_cells}
\title{Fill merged Excel cells down}
\usage{
fill_merged_cells(...)
}
\arguments{
\item{...}{One or more Excel sheets with merged cells}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function fills merged Excel cells down
}
