% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_files}
\alias{find_files}
\title{Find files stored in a directory according to a naming pattern}
\usage{
find_files(.path, .ext = ".csv")
}
\arguments{
\item{.path}{directory tree}

\item{.ext}{file extension}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function finds files stored in a directory according to a naming pattern
}
