% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_digits}
\alias{format_digits}
\title{Round and set a minimum number of digits}
\usage{
format_digits(.var, .digits)
}
\arguments{
\item{.var}{a numeric vector}

\item{.digits}{the minimum number of digits that will be set}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function round and set a minimum number of digits
}
