% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_colnames}
\alias{get_colnames}
\title{Get the name of each variable in a data frame}
\usage{
get_colnames(.data)
}
\arguments{
\item{.data}{data frame}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function gests the name of each variable in a data frame
}
