% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Get duplicate rows in a data frame}
\usage{
get_duplicates(.data)
}
\arguments{
\item{.data}{data frame}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function gets duplicate rows in a data frame
}
\examples{
data.frame(
   var1 = c(1,1,1,1),
   var2 = c(2,2,2,3)
) -> data

 get_duplicates(data)
}
