% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_files}
\alias{list_files}
\title{List files stored in a directory according to a naming pattern}
\usage{
list_files(.path, .ext = ".csv")
}
\arguments{
\item{.path}{directory tree}

\item{.ext}{file extension}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function lists files stored in a directory according to a naming pattern
}
