% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_select}
\alias{list_select}
\title{Select objects in lists}
\usage{
list_select(.list, ...)
}
\arguments{
\item{.list}{a list}

\item{...}{objects}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function selects named or unnamed objects in lists. It is equivalent to
\code{list[c("object")]} or \code{list[c(1,2)]}
}
