% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_all_character}
\alias{make_all_character}
\title{Make all data frame variables character}
\usage{
make_all_character(...)
}
\arguments{
\item{...}{A single data frame or a list of data frames}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function makes all data frame variables character
}
