% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_call}
\alias{match_call}
\title{Argument matching with formals}
\usage{
match_call()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function returns a list of a function's arguments and their values, including the default values
}
\examples{
f = function(x = 1, y = 2) match_call()
f(x = 3)
f = function(x = 1, y = 2, ...) match_call()
f(x = 3, dots = 13)
}
\references{
https://stackoverflow.com/questions/14397364/match-call-with-default-arguments
}
