% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{not_in}
\alias{not_in}
\title{Are the elements of x contained in y?}
\usage{
not_in(x, y)
}
\arguments{
\item{x}{an \R object to be tested}

\item{y}{an \R object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function indicates which elements of x are not in y
}
