% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_delim_zip}
\alias{read_delim_zip}
\title{Read delimited files stored in a zip file}
\usage{
read_delim_zip(x, encoding = "latin1", delim = ";", pattern = "", ...)
}
\arguments{
\item{x}{zip file.}

\item{delim}{See \code{\link[vroom:vroom]{vroom::vroom()}}}

\item{pattern}{A regular expression to be matched in the given compressed files character vector.}

\item{...}{\code{\link[vroom:vroom]{vroom::vroom()}} args.}

\item{enconding}{Set the encoding of the delimited files.}
}
\description{
This function reads delimited files stored in a zip file
}
