% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_all}
\alias{rename_all}
\title{Rename columns in a list of dataframes}
\usage{
rename_all(list, ...)
}
\arguments{
\item{list}{list of data frames}

\item{...}{new_name = "old_name"}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function renames columns in a list of dataframes.
}
