% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_quarto}
\alias{render_quarto}
\title{Render quarto documents in pdf, html and docx formats}
\usage{
render_quarto(.quarto_doc, .params)
}
\arguments{
\item{.quarto_doc}{Quarto document that will be rendered}

\item{.params}{A list of named parameters}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function renders quarto documents in pdf, html and docx formats
}
