% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_group.R
\name{split_group}
\alias{split_group}
\title{Split data frame by groups}
\usage{
split_group(data, ...)
}
\arguments{
\item{data}{A dataframe}

\item{...}{group variables}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function splits data frames by groups and returns a list of data frames
}
\examples{
split_group(mtcars, vs, am)
}
\references{
https://github.com/markfairbanks/tidytable
}
