% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_data}
\alias{update_data}
\title{Update the raw data of a package}
\usage{
update_data(.path = "data-raw", .from_scratch, ...)
}
\arguments{
\item{.path}{path}

\item{.from_scratch}{reads each R script inside \code{.path}}

\item{...}{data}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function updates the raw data of a package
}
