% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_package}
\alias{update_package}
\title{Update documentation, Rmarkdown files and raw data; check, test,
build and reinstall the running package locally.}
\usage{
update_package(
  .update_data = FALSE,
  .update_readme = FALSE,
  .update_rmd = FALSE,
  .update_site = FALSE,
  .check = FALSE,
  .unit_test = FALSE
)
}
\arguments{
\item{.update_data}{Update raw data}

\item{.update_readme}{Update the README file}

\item{.update_rmd}{Build Rmarkdown files inside the package}

\item{.update_site}{Build the pkgdown website}

\item{.check}{Check the package}

\item{.unit_test}{Execute \code{testthat} tests}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function updates documentation, Rmarkdown files and raw data; check, test,
build and reinstall the running package locally.
}
