% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_var}
\alias{check_var}
\title{Check if a variable exists within a data frame}
\usage{
check_var(.data, .var)
}
\arguments{
\item{.data}{data frame}

\item{.var}{variable on which existence will be checked}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function checks if a variable exists within a data frame
}
