% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtools-package.R
\docType{package}
\name{dtools-package}
\alias{dtools}
\alias{dtools-package}
\title{dtools: Tools for data analysis}
\description{
This package provides tools for data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arturhgq/dtools}
  \item Report bugs at \url{https://github.com/arturhgq/dtools/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
